Tidy up `ResultsScreen`
Just the bare minimum code quality so I can start working on these
classes..

Please push back if this doesn't seem better than what was already
there. This is mostly autopilot fixing for me based on how I've been
writing code for osu! to date.

There are changes to the load process but nothing which should cause
issues, I hope.
